Create Ergonomics base class
