+++ api/src/main/java/io/hyperfoil/api/config/SLABuilder.java
@@ -24,6 +24,7 @@ public class SLABuilder<P> implements BuilderBase<SLABuilder<P>> {
       this.parent = parent;
    }
 
+   @Override
    public void prepareBuild() {
    }
